feat: new file structure, better logging and slight changes in the test pipeline
